requirements.txt
pandas
psycopg2-binary
gradio
python-dotenv
openai
tldextract
bs4